+++ src/app_pages/live_map.py
@@ -11,6 +11,7 @@ from utils import gtfs_static
 from utils import route_view
 from utils import walking
 from utils import background_fetch
+from app_pages.live_map_components import deck_layers, stop_panel, route_panel
 
 try:
     from config import DEFAULT_ZOOM, ARROW_SIZE